+++ src/py21cmmclite/likelihood.py
@@ -1,6 +1,7 @@
 from .coeval import CoevalNeutralFraction, EoRSimulator
 import numpy as np
 import py21cmfast as p21
+from .lightcone import LightconeNeutralFraction
 
 
 def compute_chi2(model_vector, data_vector, inv_covariance):
@@ -125,7 +126,7 @@ class LikelihoodGaussian(LikelihoodBase):
 
 class LikelihoodCoevalNeutralFraction(EoRSimulator, LikelihoodGaussian):
     """
-    Likelihood for neutral fraction.
+    Likelihood for neutral fraction for coeval cubes.
     """
 
     def __init__(
@@ -160,5 +161,58 @@ class LikelihoodCoevalNeutralFraction(EoRSimulator, LikelihoodGaussian):
                 global_params=global_params,
                 save_global_xhi=save_global_xhi,
                 save_xhi_box=save_xhi_box,
+            )
+        ]
+
+
+class LikelihoodLightconeNeutralFraction(EoRSimulator, LikelihoodGaussian):
+    """
+    Likelihood for neutral fraction for lightcone at different redshift bins.
+    """
+
+    def __init__(
+        self,
+        xhi_z_edges_low: list[float],
+        xhi_z_edges_high: list[float],
+        inputs_21cmfast: p21.InputParameters,
+        varied_params: list[str],
+        cache_dir: str,
+        regenerate: bool = False,
+        global_params: dict | None = None,
+        lc_min_redshift: float | None = None,
+        lc_max_redshift: float | None = None,
+        lc_quantities: list[str] = ["brightness_temp", "neutral_fraction"],
+        data_dict: dict | None = None,
+        simulate_data: bool = False,
+        simulate_error_fraction: float = 0.1,
+        save_xhi_points: bool = False,
+        save_xhi_lc: bool = False,
+    ):
+        EoRSimulator.__init__(
+            self, inputs_21cmfast, cache_dir, regenerate, global_params
+        )
+        LikelihoodGaussian.__init__(
+            self, varied_params, data_dict, simulate_data, simulate_error_fraction
+        )
+        self.xhi_z_edges_low = xhi_z_edges_low
+        self.xhi_z_edges_high = xhi_z_edges_high
+        self.lc_min_redshift = lc_min_redshift
+        self.lc_max_redshift = lc_max_redshift
+        self.lc_quantities = lc_quantities
+        self.save_xhi_points = save_xhi_points
+        self.save_xhi_lc = save_xhi_lc
+        self.simulators = [
+            LightconeNeutralFraction(
+                xhi_z_edges_low=xhi_z_edges_low,
+                xhi_z_edges_high=xhi_z_edges_high,
+                inputs=inputs_21cmfast,
+                cache_dir=cache_dir,
+                regenerate=regenerate,
+                global_params=global_params,
+                lc_min_redshift=lc_min_redshift,
+                lc_max_redshift=lc_max_redshift,
+                lc_quantities=lc_quantities,
+                save_xhi_points=save_xhi_points,
+                save_xhi_lc=save_xhi_lc,
             )
         ]