+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Added
 
 - Add MPI support for emcee and nautilus
 - A simple lightcone simulator that can be used to simulate the lightcone and calculate the global neutral fraction
+- Allow likelihood class to invoke the lightcone simulator
 
 Fixed
 ~~~~~